Add base page structure, font style and README.md
The main HTML is now set and will be used as a
skeleton for building the landing page, I have also
added the font that I will be using ('Roboto') and
have written a basic description for the project in the
README file.

I plan to be as loyal as possible to the main reference
design, and then customize the page the way I like and
give it my own flair when finished.