Add second section + styling
The second section has a lower padding amount
than the first. Inside nested flex containers I
added the section title, as well as the container that holds
the images and their subtexts, both on the same nested flex container.

There's probably gonna be some problems
when trying to add the images, but that's okay, as long
as the base is working well.